+++ infra/resources/_modules/function_app_sending/locals.tf
@@ -23,7 +23,7 @@ locals {
       NEW_NOTIFICATION_QUEUE_STORAGE_CONNECTION_STRING = var.com_st_connectiostring
       NEW_NOTIFICATION_QUEUE_NAME                      = "push-notifications"
 
-      IO_COM_PUSH_NOTIFICATIONS_REDIRECT_PERCENTAGE = "0"
+      IO_COM_PUSH_NOTIFICATIONS_REDIRECT_PERCENTAGE = "1"
 
       // REDIS
       REDIS_URL      = var.redis_url
@@ -33,6 +33,9 @@ locals {
       // BACKEND COMMUNICATION
       BACKEND_BASE_URL = "https://io-p-app-appbackendli.azurewebsites.net"
       BACKEND_TOKEN    = var.appbackendli_token
+
+      SESSION_MANAGER_API_KEY  = "@Microsoft.KeyVault(VaultName=${var.com_kv_name};SecretName=session-manager-api-key)",
+      SESSION_MANAGER_BASE_URL = var.session_manager_base_url,
 
       // INTERNAL USE PROPERTIES
       INTERNAL_USER_ID           = var.internal_user_id